Moved some things around
