+++ src/app.py
@@ -1,4 +1,26 @@
-from utils import db_connect
-engine = db_connect()
+from flask import Flask, render_template, request
+import pickle
 
-# your code here
+app = Flask(__name__)
+cv = pickle.load(open("models/cv.pkl", 'rb'))
+clf = pickle.load(open("models/clf.pkl", 'rb'))
+
+
+@app.route('/')
+def index():
+    return render_template("index.html")
+
+
+@app.route('/predict', methods=['post'])
+def predict():
+    email = request.form.get('email')
+    # predict email
+    print(email)
+    X = cv.transform([email])
+    prediction = clf.predict(X)
+    prediction = 1 if prediction == 1 else -1
+    return render_template('index.html', response=prediction)
+
+
+if __name__ == "__main__":
+    app.run(debug=True)
